New Guard module for stopping execution when conditions aren't met. Honey __before_compile__ is now concise.
